+++ apps/meteor/client/views/room/body/RoomBody.tsx
@@ -1,24 +1,17 @@
-import type { IMessage, IUser } from '@rocket.chat/core-typings';
-import { isEditedMessage } from '@rocket.chat/core-typings';
-import { Box, Bubble } from '@rocket.chat/fuselage';
+import type { IUser } from '@rocket.chat/core-typings';
+import { Box } from '@rocket.chat/fuselage';
 import { useMergedRefs } from '@rocket.chat/fuselage-hooks';
-import { usePermission, useRole, useRouter, useSetting, useTranslation, useUser, useUserPreference } from '@rocket.chat/ui-contexts';
+import { usePermission, useRole, useSetting, useTranslation, useUser, useUserPreference } from '@rocket.chat/ui-contexts';
 import type { MouseEventHandler, ReactElement, UIEvent } from 'react';
-import React, { memo, useCallback, useEffect, useMemo, useRef, useState } from 'react';
+import React, { memo, useCallback, useMemo, useRef } from 'react';
 
-import { ChatMessage, RoomRoles } from '../../../../app/models/client';
-import { RoomHistoryManager } from '../../../../app/ui-utils/client';
-import { isAtBottom } from '../../../../app/ui/client/views/app/lib/scrolling';
-import { callbacks } from '../../../../lib/callbacks';
+import { RoomRoles } from '../../../../app/models/client';
 import { isTruthy } from '../../../../lib/isTruthy';
-import { withDebouncing, withThrottling } from '../../../../lib/utils/highOrderFunctions';
 import { CustomScrollbars } from '../../../components/CustomScrollbars';
 import { useEmbeddedLayout } from '../../../hooks/useEmbeddedLayout';
-import { useFormatDate } from '../../../hooks/useFormatDate';
 import { useReactiveQuery } from '../../../hooks/useReactiveQuery';
-import { roomCoordinator } from '../../../lib/rooms/roomCoordinator';
-import { setMessageJumpQueryStringParameter } from '../../../lib/utils/setMessageJumpQueryStringParameter';
 import Announcement from '../Announcement';
+import { BubbleDate } from '../BubbleDate';
 import { MessageList } from '../MessageList';
 import MessageListErrorBoundary from '../MessageList/MessageListErrorBoundary';
 import ComposerContainer from '../composer/ComposerContainer';
@@ -29,7 +22,6 @@ import { useRoomToolbox } from '../contexts/RoomToolboxContext';
 import { useUserCard } from '../contexts/UserCardContext';
 import { useDateScroll } from '../hooks/useDateScroll';
 import { useMessageListNavigation } from '../hooks/useMessageListNavigation';
-import { useDateListController } from '../providers/DateListProvider';
 import DropTargetOverlay from './DropTargetOverlay';
 import JumpToRecentMessageButton from './JumpToRecentMessageButton';
 import LeaderBar from './LeaderBar';
@@ -39,16 +31,23 @@ import RoomForeword from './RoomForeword/RoomForeword';
 import UnreadMessagesIndicator from './UnreadMessagesIndicator';
 import UploadProgressIndicator from './UploadProgressIndicator';
 import { useFileUpload } from './hooks/useFileUpload';
+import { useGetMore } from './hooks/useGetMore';
 import { useGoToHomeOnRemoved } from './hooks/useGoToHomeOnRemoved';
+import { useHasNewMessages } from './hooks/useHasNewMessages';
+import { useLeaderBanner } from './hooks/useLeaderBanner';
 import { useListIsAtBottom } from './hooks/useListIsAtBottom';
+import { useQuoteMessageByUrl } from './hooks/useQuoteMessageByUrl';
 import { useReadMessageWindowEvents } from './hooks/useReadMessageWindowEvents';
 import { useRestoreScrollPosition } from './hooks/useRestoreScrollPosition';
 import { useRetentionPolicy } from './hooks/useRetentionPolicy';
-import { useUnreadMessages } from './hooks/useUnreadMessages';
-import { useQuoteMessageByUrl } from './hooks/useQuoteMessageByUrl';
+import { useHandleUnread } from './hooks/useUnreadMessages';
 
 const RoomBody = (): ReactElement => {
-	const formatDate = useFormatDate();
+	const chat = useChat();
+	if (!chat) {
+		throw new Error('No ChatContext provided');
+	}
+
 	const t = useTranslation();
 	const isLayoutEmbedded = useEmbeddedLayout();
 	const room = useRoom();
@@ -57,75 +56,11 @@ const RoomBody = (): ReactElement => {
 	const admin = useRole('admin');
 	const subscription = useRoomSubscription();
 
-	const { list } = useDateListController();
-	const { callbackRef, listStyle, bubbleDate, showBubble, style: bubbleDateStyle } = useDateScroll();
-
-	const [lastMessageDate, setLastMessageDate] = useState<Date | undefined>();
-	const [hideLeaderHeader, setHideLeaderHeader] = useState(false);
-	const [hasNewMessages, setHasNewMessages] = useState(false);
+	const retentionPolicy = useRetentionPolicy(room);
 
 	const hideFlexTab = useUserPreference<boolean>('hideFlexTab') || undefined;
 	const hideUsernames = useUserPreference<boolean>('hideUsernames');
 	const displayAvatars = useUserPreference<boolean>('displayAvatars');
-
-	const wrapperRef = useRef<HTMLDivElement | null>(null);
-	const messagesBoxRef = useRef<HTMLDivElement | null>(null);
-
-	const { atBottomRef, ref: isAtBottomCallbackRef } = useListIsAtBottom();
-
-	const chat = useChat();
-	const { openUserCard, triggerProps } = useUserCard();
-
-	if (!chat) {
-		throw new Error('No ChatContext provided');
-	}
-
-	const {
-		uploads,
-		handleUploadFiles,
-		handleUploadProgressClose,
-		targeDrop: [fileUploadTriggerProps, fileUploadOverlayProps],
-	} = useFileUpload();
-
-	const _isAtBottom = useCallback((scrollThreshold = 0) => {
-		const wrapper = wrapperRef.current;
-
-		if (!wrapper) {
-			return false;
-		}
-
-		if (isAtBottom(wrapper, scrollThreshold)) {
-			setHasNewMessages(false);
-			return true;
-		}
-		return false;
-	}, []);
-
-	const sendToBottom = useCallback(() => {
-		wrapperRef.current?.scrollTo({ left: 30, top: wrapperRef.current?.scrollHeight });
-
-		setHasNewMessages(false);
-	}, []);
-
-	const sendToBottomIfNecessary = useCallback(() => {
-		if (atBottomRef.current === true) {
-			sendToBottom();
-		}
-	}, [atBottomRef, sendToBottom]);
-
-	const handleNewMessageButtonClick = useCallback(() => {
-		atBottomRef.current = true;
-		sendToBottomIfNecessary();
-		chat.composer?.focus();
-	}, [atBottomRef, chat.composer, sendToBottomIfNecessary]);
-
-	const handleJumpToRecentButtonClick = useCallback(() => {
-		atBottomRef.current = true;
-		RoomHistoryManager.clear(room._id);
-		RoomHistoryManager.getMoreIfIsEmpty(room._id);
-	}, [atBottomRef, room._id]);
-
-	const [unread, setUnreadCount] = useUnreadMessages(room);
 
 	const { hasMorePreviousMessages, hasMoreNextMessages, isLoadingMoreMessages } = useRoomMessages();
 
@@ -153,232 +88,56 @@ const RoomBody = (): ReactElement => {
 
 	const useRealName = useSetting('UI_Use_Real_Name') as boolean;
 
-	const { data: roomLeader } = useReactiveQuery(['rooms', room._id, 'leader', { not: user?._id }], () => {
-		const leaderRoomRole = RoomRoles.findOne({
-			'rid': room._id,
-			'roles': 'leader',
-			'u._id': { $ne: user?._id },
+	const innerBoxRef = useRef<HTMLDivElement | null>(null);
+
+	const {
+		wrapperRef: unreadBarWrapperRef,
+		innerRef: unreadBarInnerRef,
+		handleUnreadBarJumpToButtonClick,
+		handleMarkAsReadButtonClick,
+		counter: [unread],
+	} = useHandleUnread(room, subscription);
+
+	const { innerRef: dateScrollInnerRef, bubbleRef, listStyle, ...bubbleDate } = useDateScroll();
+
+	const { innerRef: isAtBottomInnerRef, atBottomRef, sendToBottom, sendToBottomIfNecessary, isAtBottom } = useListIsAtBottom();
+
+	const { innerRef: getMoreInnerRef } = useGetMore(room._id, atBottomRef);
+
+	const { wrapperRef: leaderBannerWrapperRef, hideLeaderHeader, innerRef: leaderBannerInnerRef } = useLeaderBanner();
+
+	const {
+		uploads,
+		handleUploadFiles,
+		handleUploadProgressClose,
+		targeDrop: [fileUploadTriggerProps, fileUploadOverlayProps],
+	} = useFileUpload();
+
+	const { innerRef: restoreScrollPositionInnerRef } = useRestoreScrollPosition(room._id);
+
+	const { messageListRef, messageListProps } = useMessageListNavigation();
+
+	const { handleNewMessageButtonClick, handleJumpToRecentButtonClick, handleComposerResize, hasNewMessages, newMessagesScrollRef } =
+		useHasNewMessages(room._id, user?._id, atBottomRef, {
+			sendToBottom,
+			sendToBottomIfNecessary,
+			isAtBottom,
 		});
 
-		if (!leaderRoomRole) {
-			return null;
-		}
+	const innerRef = useMergedRefs(
+		dateScrollInnerRef,
+		innerBoxRef,
+		restoreScrollPositionInnerRef,
+		isAtBottomInnerRef,
+		newMessagesScrollRef,
+		leaderBannerInnerRef,
+		unreadBarInnerRef,
+		getMoreInnerRef,
 
-		return {
-			...leaderRoomRole.u,
-			name: useRealName ? leaderRoomRole.u.name || leaderRoomRole.u.username : leaderRoomRole.u.username,
-		};
-	});
-
-	const handleOpenUserCard = useCallback(
-		(event: UIEvent, username: IUser['username']) => {
-			if (!username) {
-				return;
-			}
-
-			openUserCard(event, username);
-		},
-		[openUserCard],
+		messageListRef,
 	);
 
-	const handleUnreadBarJumpToButtonClick = useCallback(() => {
-		const rid = room._id;
-		const { firstUnread } = RoomHistoryManager.getRoom(rid);
-		let message = firstUnread?.get();
-		if (!message) {
-			message = ChatMessage.findOne({ rid, ts: { $gt: unread?.since } }, { sort: { ts: 1 }, limit: 1 });
-		}
-		if (!message) {
-			return;
-		}
-		setMessageJumpQueryStringParameter(message?._id);
-		setUnreadCount(0);
-	}, [room._id, unread?.since, setUnreadCount]);
-
-	const handleMarkAsReadButtonClick = useCallback(() => {
-		chat.readStateManager.markAsRead();
-		setUnreadCount(0);
-	}, [chat.readStateManager, setUnreadCount]);
-
-	const retentionPolicy = useRetentionPolicy(room);
-
-	useGoToHomeOnRemoved(room, user?._id);
-
-	useEffect(() => {
-		callbacks.add(
-			'streamNewMessage',
-			(msg: IMessage) => {
-				if (room._id !== msg.rid || isEditedMessage(msg) || msg.tmid) {
-					return;
-				}
-
-				if (msg.u._id === user?._id) {
-					sendToBottom();
-					return;
-				}
-
-				if (!_isAtBottom()) {
-					setHasNewMessages(true);
-				}
-			},
-			callbacks.priority.MEDIUM,
-			room._id,
-		);
-
-		return () => {
-			callbacks.remove('streamNewMessage', room._id);
-		};
-	}, [_isAtBottom, room._id, sendToBottom, user?._id]);
-
-	useEffect(() => {
-		const messageList = wrapperRef.current?.querySelector('ul');
-
-		if (!messageList) {
-			return;
-		}
-
-		const observer = new ResizeObserver(() => {
-			sendToBottomIfNecessary();
-		});
-
-		observer.observe(messageList);
-
-		return () => {
-			observer?.disconnect();
-		};
-	}, [sendToBottomIfNecessary]);
-
-	const router = useRouter();
-
-	const debouncedReadMessageRead = useMemo(
-		() =>
-			withDebouncing({ wait: 500 })(() => {
-				chat.readStateManager.attemptMarkAsRead();
-			}),
-		[chat.readStateManager],
-	);
-
-	useEffect(
-		() =>
-			router.subscribeToRouteChange(() => {
-				const routeName = router.getRouteName();
-				if (!routeName || !roomCoordinator.isRouteNameKnown(routeName)) {
-					return;
-				}
-
-				debouncedReadMessageRead();
-			}),
-		[debouncedReadMessageRead, room._id, router, subscribed, subscription?.alert, subscription?.unread],
-	);
-
-	useEffect(() => {
-		if (!subscribed) {
-			setUnreadCount(0);
-			return;
-		}
-
-		const count = ChatMessage.find({
-			rid: room._id,
-			ts: { $lte: lastMessageDate, $gt: subscription?.ls },
-		}).count();
-
-		setUnreadCount(count);
-	}, [lastMessageDate, room._id, setUnreadCount, subscribed, subscription?.ls]);
-
-	useEffect(() => {
-		if (!unread?.count) {
-			return debouncedReadMessageRead();
-		}
-	}, [debouncedReadMessageRead, room._id, unread?.count]);
-
-	useEffect(() => {
-		const wrapper = wrapperRef.current;
-
-		if (!wrapper) {
-			return;
-		}
-
-		const getElementFromPoint = (topOffset = 0): Element | undefined => {
-			const messagesBox = messagesBoxRef.current;
-
-			if (!messagesBox) {
-				return;
-			}
-
-			const messagesBoxLeft = messagesBox.getBoundingClientRect().left + window.pageXOffset;
-			const messagesBoxTop = messagesBox.getBoundingClientRect().top + window.pageYOffset;
-			const messagesBoxWidth = parseFloat(getComputedStyle(messagesBox).width);
-
-			let element;
-			if (document.dir === 'rtl') {
-				element = document.elementFromPoint(messagesBoxLeft + messagesBoxWidth - 1, messagesBoxTop + topOffset + 1);
-			} else {
-				element = document.elementFromPoint(messagesBoxLeft + 1, messagesBoxTop + topOffset + 1);
-			}
-
-			if (element?.classList.contains('rcx-message') || element?.classList.contains('rcx-message--sequential')) {
-				return element;
-			}
-		};
-
-		const updateUnreadCount = withThrottling({ wait: 300 })(() => {
-			Tracker.afterFlush(() => {
-				const lastInvisibleMessageOnScreen = getElementFromPoint(0) || getElementFromPoint(20) || getElementFromPoint(40);
-
-				if (!lastInvisibleMessageOnScreen) {
-					setUnreadCount(0);
-					return;
-				}
-
-				const lastMessage = ChatMessage.findOne(lastInvisibleMessageOnScreen.id);
-				if (!lastMessage) {
-					setUnreadCount(0);
-					return;
-				}
-
-				setLastMessageDate(lastMessage.ts);
-			});
-		});
-
-		let lastScrollTopRef = 0;
-		const handleWrapperScroll = withThrottling({ wait: 100 })((event) => {
-			const roomLeader = messagesBoxRef.current?.querySelector('.room-leader');
-			if (roomLeader) {
-				if (event.target.scrollTop < lastScrollTopRef) {
-					setHideLeaderHeader(false);
-				} else if (_isAtBottom(100) === false && event.target.scrollTop > parseFloat(getComputedStyle(roomLeader).height)) {
-					setHideLeaderHeader(true);
-				}
-			}
-			lastScrollTopRef = event.target.scrollTop;
-			const height = event.target.clientHeight;
-			const isLoading = RoomHistoryManager.isLoading(room._id);
-			const hasMore = RoomHistoryManager.hasMore(room._id);
-			const hasMoreNext = RoomHistoryManager.hasMoreNext(room._id);
-
-			if ((isLoading === false && hasMore === true) || hasMoreNext === true) {
-				if (hasMore === true && lastScrollTopRef <= height / 3) {
-					RoomHistoryManager.getMore(room._id);
-				} else if (hasMoreNext === true && Math.ceil(lastScrollTopRef) >= event.target.scrollHeight - height) {
-					RoomHistoryManager.getMoreNext(room._id, atBottomRef);
-				}
-			}
-		});
-
-		wrapper.addEventListener('scroll', updateUnreadCount);
-		wrapper.addEventListener('scroll', handleWrapperScroll);
-
-		return () => {
-			wrapper.removeEventListener('scroll', updateUnreadCount);
-			wrapper.removeEventListener('scroll', handleWrapperScroll);
-		};
-	}, [_isAtBottom, atBottomRef, list, room._id, setUnreadCount]);
-
-	const restoreScrollPositionRef = useRestoreScrollPosition(room._id);
-
-	const handleComposerResize = useCallback((): void => {
-		sendToBottomIfNecessary();
-	}, [sendToBottomIfNecessary]);
+	const wrapperBoxRefs = useMergedRefs(unreadBarWrapperRef, leaderBannerWrapperRef);
 
 	const handleNavigateToPreviousMessage = useCallback((): void => {
 		chat.messageEditing.toPreviousMessage();
@@ -387,10 +146,6 @@ const RoomBody = (): ReactElement => {
 	const handleNavigateToNextMessage = useCallback((): void => {
 		chat.messageEditing.toNextMessage();
 	}, [chat.messageEditing]);
-
-	useEffect(() => {
-		chat.uploads.wipeFailedOnes();
-	}, [chat]);
 
 	const handleCloseFlexTab: MouseEventHandler<HTMLElement> = useCallback(
 		(e): void => {
@@ -424,55 +179,87 @@ const RoomBody = (): ReactElement => {
 		[toolbox],
 	);
 
+	const { openUserCard, triggerProps } = useUserCard();
+
+	const handleOpenUserCard = useCallback(
+		(event: UIEvent, username: IUser['username']) => {
+			if (!username) {
+				return;
+			}
+
+			openUserCard(event, username);
+		},
+		[openUserCard],
+	);
+
+	useGoToHomeOnRemoved(room, user?._id);
 	useReadMessageWindowEvents();
 	useQuoteMessageByUrl();
 
-	const { messageListRef, messageListProps } = useMessageListNavigation();
+	const { data: roomLeader } = useReactiveQuery(['rooms', room._id, 'leader', { not: user?._id }], () => {
+		const leaderRoomRole = RoomRoles.findOne({
+			'rid': room._id,
+			'roles': 'leader',
+			'u._id': { $ne: user?._id },
+		});
 
-	const ref = useMergedRefs(callbackRef, wrapperRef, restoreScrollPositionRef, isAtBottomCallbackRef);
+		if (!leaderRoomRole) {
+			return null;
+		}
+
+		return {
+			...leaderRoomRole.u,
+			name: useRealName ? leaderRoomRole.u.name || leaderRoomRole.u.username : leaderRoomRole.u.username,
+		};
+	});
 
 	return (
 		<>
 			{!isLayoutEmbedded && room.announcement && <Announcement announcement={room.announcement} announcementDetails={undefined} />}
-			<Box className={['main-content-flex', listStyle]}>
+			<Box key={room._id} className={['main-content-flex', listStyle]}>
 				<section
 					className={`messages-container flex-tab-main-content ${admin ? 'admin' : ''}`}
 					id={`chat-window-${room._id}`}
 					onClick={hideFlexTab && handleCloseFlexTab}
 				>
 					<div className='messages-container-wrapper'>
-						<div className='messages-container-main' ref={callbackRef} key={room._id} {...fileUploadTriggerProps}>
+						<div className='messages-container-main' ref={wrapperBoxRefs} {...fileUploadTriggerProps}>
 							<DropTargetOverlay {...fileUploadOverlayProps} />
-							<div className={['container-bars', uploads.length && 'show'].filter(isTruthy).join(' ')}>
-								{uploads.map((upload) => (
-									<UploadProgressIndicator
-										key={upload.id}
-										id={upload.id}
-										name={upload.name}
-										percentage={upload.percentage}
-										error={upload.error instanceof Error ? upload.error.message : undefined}
-										onClose={handleUploadProgressClose}
+							<Box position='absolute' w='full'>
+								{roomLeader ? (
+									<LeaderBar
+										_id={roomLeader._id}
+										username={roomLeader.username}
+										name={roomLeader.name}
+										visible={!hideLeaderHeader}
+										onAvatarClick={handleOpenUserCard}
+										triggerProps={triggerProps}
 									/>
-								))}
-							</div>
-							{bubbleDate && (
-								<Box className={[bubbleDateStyle, showBubble && 'bubble-visible']}>
-									<Bubble small secondary>
-										{formatDate(bubbleDate)}
-									</Bubble>
-								</Box>
-							)}
-							{unread && (
-								<UnreadMessagesIndicator
-									count={unread.count}
-									onJumpButtonClick={handleUnreadBarJumpToButtonClick}
-									onMarkAsReadButtonClick={handleMarkAsReadButtonClick}
-								/>
-							)}
-							<div
-								ref={messagesBoxRef}
-								className={['messages-box', roomLeader && !hideLeaderHeader && 'has-leader'].filter(isTruthy).join(' ')}
-							>
+								) : null}
+								<div className={['container-bars', uploads.length && 'show'].filter(isTruthy).join(' ')}>
+									{uploads.map((upload) => (
+										<UploadProgressIndicator
+											key={upload.id}
+											id={upload.id}
+											name={upload.name}
+											percentage={upload.percentage}
+											error={upload.error instanceof Error ? upload.error.message : undefined}
+											onClose={handleUploadProgressClose}
+										/>
+									))}
+								</div>
+								{Boolean(unread) && (
+									<UnreadMessagesIndicator
+										count={unread}
+										onJumpButtonClick={handleUnreadBarJumpToButtonClick}
+										onMarkAsReadButtonClick={handleMarkAsReadButtonClick}
+									/>
+								)}
+
+								<BubbleDate ref={bubbleRef} {...bubbleDate} />
+							</Box>
+
+							<div className={['messages-box', roomLeader && !hideLeaderHeader && 'has-leader'].filter(isTruthy).join(' ')}>
 								<JumpToRecentMessageButton visible={hasNewMessages} onClick={handleNewMessageButtonClick} text={t('New_messages')} />
 								<JumpToRecentMessageButton
 									visible={hasMoreNextMessages}
@@ -483,16 +270,6 @@ const RoomBody = (): ReactElement => {
 									<div className='content room-not-found error-color'>
 										<div>{t('You_must_join_to_view_messages_in_this_channel')}</div>
 									</div>
-								) : null}
-								{roomLeader ? (
-									<LeaderBar
-										_id={roomLeader._id}
-										username={roomLeader.username}
-										name={roomLeader.name}
-										visible={!hideLeaderHeader}
-										onAvatarClick={handleOpenUserCard}
-										triggerProps={triggerProps}
-									/>
 								) : null}
 								<div
 									className={[
@@ -505,14 +282,8 @@ const RoomBody = (): ReactElement => {
 										.join(' ')}
 								>
 									<MessageListErrorBoundary>
-										<CustomScrollbars ref={ref}>
-											<ul
-												ref={messageListRef}
-												className='messages-list'
-												aria-live='polite'
-												aria-busy={isLoadingMoreMessages}
-												{...messageListProps}
-											>
+										<CustomScrollbars ref={innerRef}>
+											<ul className='messages-list' aria-live='polite' aria-busy={isLoadingMoreMessages} {...messageListProps}>
 												{canPreview ? (
 													<>
 														{hasMorePreviousMessages ? (
@@ -525,7 +296,7 @@ const RoomBody = (): ReactElement => {
 														)}
 													</>
 												) : null}
-												<MessageList rid={room._id} messageListRef={wrapperRef} />
+												<MessageList rid={room._id} messageListRef={innerBoxRef} />
 												{hasMoreNextMessages ? (
 													<li className='load-more'>{isLoadingMoreMessages ? <LoadingMessagesIndicator /> : null}</li>
 												) : null}