+++ apps/meteor/server/services/federation/infrastructure/matrix/Bridge.ts
@@ -714,7 +714,7 @@ export class MatrixBridge implements IFederationBridge {
 		// Dynamic import to prevent Rocket.Chat from loading the module until needed and then handle if that fails
 		const { Bridge, AppServiceRegistration, MatrixUser } = await import('@rocket.chat/forked-matrix-appservice-bridge');
 		MatrixUserInstance = MatrixUser;
-		const registrationFile = this.internalSettings.generateRegistrationFileObject();
+		const registrationFile = this.internalSettings.getAppServiceRegistrationObject();
 
 		this.bridgeInstance = new Bridge({
 			homeserverUrl: this.internalSettings.getHomeServerUrl(),
@@ -724,13 +724,12 @@ export class MatrixBridge implements IFederationBridge {
 			controller: {
 				onEvent: (request) => {
 					const event = request.getData() as unknown as AbstractMatrixEvent;
-					console.log({ event });
 					this.eventHandler(event);
 				},
 				onLog: (line, isError) => {
 					console.log(line, isError);
 				},
-				...(this.internalSettings.generateRegistrationFileObject().enableEphemeralEvents
+				...(this.internalSettings.getAppServiceRegistrationObject().enableEphemeralEvents
 					? {
 							onEphemeralEvent: (request) => {
 								const event = request.getData() as unknown as AbstractMatrixEvent;