+++ apps/meteor/server/services/omnichannel/queue.ts
@@ -1,6 +1,7 @@
-import type { InquiryWithAgentInfo, IOmnichannelQueue } from '@rocket.chat/core-typings';
+import type { IOmnichannelRoom } from '@rocket.chat/core-typings';
+import { type InquiryWithAgentInfo, type IOmnichannelQueue } from '@rocket.chat/core-typings';
 import { License } from '@rocket.chat/license';
-import { LivechatInquiry } from '@rocket.chat/models';
+import { LivechatInquiry, LivechatRooms } from '@rocket.chat/models';
 
 import { dispatchAgentDelegated } from '../../../app/livechat/server/lib/Helper';
 import { RoutingManager } from '../../../app/livechat/server/lib/RoutingManager';
@@ -38,6 +39,10 @@ export class OmnichannelQueue implements IOmnichannelQueue {
 	}
 
 	async stop() {
+		if (!this.running) {
+			return;
+		}
+
 		await LivechatInquiry.unlockAll();
 
 		this.running = false;
@@ -74,7 +79,9 @@ export class OmnichannelQueue implements IOmnichannelQueue {
 		const queueDelayTimeout = this.delay();
 		queueLogger.debug(`Executing queue ${queue || 'Public'} with timeout of ${queueDelayTimeout}`);
 
-		setTimeout(this.checkQueue.bind(this, queue), queueDelayTimeout);
+		void this.checkQueue(queue).catch((e) => {
+			queueLogger.error(e);
+		});
 	}
 
 	private async checkQueue(queue: string | undefined) {
@@ -92,10 +99,18 @@ export class OmnichannelQueue implements IOmnichannelQueue {
 				// Note: this removes the "one-shot" behavior of queue, allowing it to take a conversation again in the future
 				// And sorting them by _updatedAt: -1 will make it so that the oldest inquiries are taken first
 				// preventing us from playing with the same inquiry over and over again
+				queueLogger.debug(`Inquiry ${nextInquiry._id} not taken. Unlocking and re-queueing`);
 				return await LivechatInquiry.unlockAndQueue(nextInquiry._id);
 			}
 
+			queueLogger.debug(`Inquiry ${nextInquiry._id} taken successfully. Unlocking`);
 			await LivechatInquiry.unlock(nextInquiry._id);
+			queueLogger.debug({
+				msg: 'Inquiry processed',
+				inquiry: nextInquiry._id,
+				queue: queue || 'Public',
+				result,
+			});
 		} catch (e) {
 			queueLogger.error({
 				msg: 'Error processing queue',
@@ -103,7 +118,7 @@ export class OmnichannelQueue implements IOmnichannelQueue {
 				err: e,
 			});
 		} finally {
-			void this.execute();
+			setTimeout(this.execute.bind(this), this.delay());
 		}
 	}
 
@@ -116,24 +131,81 @@ export class OmnichannelQueue implements IOmnichannelQueue {
 		const routingSupportsAutoAssign = RoutingManager.getConfig()?.autoAssignAgent;
 		queueLogger.debug({
 			msg: 'Routing method supports auto assignment',
-			method: RoutingManager.methodName,
+			method: settings.get('Livechat_Routing_Method'),
 			status: routingSupportsAutoAssign ? 'Starting' : 'Stopping',
 		});
 
 		void (routingSupportsAutoAssign ? this.start() : this.stop());
 	}
 
+	private async reconciliation(reason: 'closed' | 'taken' | 'missing', { roomId, inquiryId }: { roomId: string; inquiryId: string }) {
+		switch (reason) {
+			case 'closed': {
+				queueLogger.debug({
+					msg: 'Room closed. Removing inquiry',
+					roomId,
+					inquiryId,
+					step: 'reconciliation',
+				});
+				await LivechatInquiry.removeByRoomId(roomId);
+				break;
+			}
+			case 'taken': {
+				queueLogger.debug({
+					msg: 'Room taken. Updating inquiry status',
+					roomId,
+					inquiryId,
+					step: 'reconciliation',
+				});
+				// Reconciliate served inquiries, by updating their status to taken after queue tried to pick and failed
+				await LivechatInquiry.takeInquiry(inquiryId);
+				break;
+			}
+			case 'missing': {
+				queueLogger.debug({
+					msg: 'Room from inquiry missing. Removing inquiry',
+					roomId,
+					inquiryId,
+					step: 'reconciliation',
+				});
+				await LivechatInquiry.removeByRoomId(roomId);
+				break;
+			}
+			default: {
+				return true;
+			}
+		}
+
+		return true;
+	}
+
 	private async processWaitingQueue(department: string | undefined, inquiry: InquiryWithAgentInfo) {
 		const queue = department || 'Public';
-		queueLogger.debug(`Processing items on queue ${queue}`);
 
 		queueLogger.debug(`Processing inquiry ${inquiry._id} from queue ${queue}`);
 		const { defaultAgent } = inquiry;
-		const room = await RoutingManager.delegateInquiry(inquiry, defaultAgent);
 
-		const propagateAgentDelegated = async (rid: string, agentId: string) => {
-			await dispatchAgentDelegated(rid, agentId);
-		};
+		const roomFromDb = await LivechatRooms.findOneById<Pick<IOmnichannelRoom, '_id' | 'servedBy' | 'closedAt'>>(inquiry.rid, {
+			projection: { servedBy: 1, closedAt: 1 },
+		});
+
+		// This is a precaution to avoid taking inquiries tied to rooms that no longer exist.
+		// This should never happen.
+		if (!roomFromDb) {
+			return this.reconciliation('missing', { roomId: inquiry.rid, inquiryId: inquiry._id });
+		}
+
+		// This is a precaution to avoid taking the same inquiry multiple times. It should not happen, but it's a safety net
+		if (roomFromDb.servedBy) {
+			return this.reconciliation('taken', { roomId: inquiry.rid, inquiryId: inquiry._id });
+		}
+
+		// This is another precaution. If the room is closed, we should not take it
+		if (roomFromDb.closedAt) {
+			return this.reconciliation('closed', { roomId: inquiry.rid, inquiryId: inquiry._id });
+		}
+
+		const room = await RoutingManager.delegateInquiry(inquiry, defaultAgent);
 
 		if (room?.servedBy) {
 			const {
@@ -142,13 +214,12 @@ export class OmnichannelQueue implements IOmnichannelQueue {
 			} = room;
 			queueLogger.debug(`Inquiry ${inquiry._id} taken successfully by agent ${agentId}. Notifying`);
 			setTimeout(() => {
-				void propagateAgentDelegated(rid, agentId);
+				void dispatchAgentDelegated(rid, agentId);
 			}, 1000);
 
 			return true;
 		}
 
-		queueLogger.debug(`Inquiry ${inquiry._id} not taken by any agent. Queueing again`);
 		return false;
 	}
 }